re-organized code
